+++ packages/admin/admin/src/form/Autocomplete.tsx
@@ -1,8 +1,21 @@
+import { ChevronDown } from "@comet/admin-icons";
 import { Autocomplete, AutocompleteProps, AutocompleteRenderInputParams, CircularProgress, InputBase } from "@mui/material";
 import * as React from "react";
 import { FieldRenderProps } from "react-final-form";
 
+import { ClearInputAdornment } from "../common/ClearInputAdornment";
 import { AsyncOptionsProps } from "../hooks/useAsyncOptionsProps";
+
+export type FinalFormAutocompleteProps<
+    T extends Record<string, any>,
+    Multiple extends boolean | undefined,
+    DisableClearable extends boolean | undefined,
+    FreeSolo extends boolean | undefined,
+> = FieldRenderProps<T, HTMLInputElement | HTMLTextAreaElement> &
+    Partial<AsyncOptionsProps<T>> &
+    Omit<AutocompleteProps<T, Multiple, DisableClearable, FreeSolo>, "renderInput"> & {
+        clearable?: boolean;
+    };
 
 export const FinalFormAutocomplete = <
     T extends Record<string, any>,
@@ -13,12 +26,14 @@ export const FinalFormAutocomplete = <
     input: { onChange, value, ...restInput },
     loading = false,
     isAsync = false,
+    clearable,
+    popupIcon = <ChevronDown />,
     ...rest
-}: FieldRenderProps<T, HTMLInputElement | HTMLTextAreaElement> &
-    Partial<AsyncOptionsProps<T>> &
-    Omit<AutocompleteProps<T, Multiple, DisableClearable, FreeSolo>, "renderInput">) => {
+}: FinalFormAutocompleteProps<T, Multiple, DisableClearable, FreeSolo>) => {
     return (
         <Autocomplete
+            popupIcon={popupIcon}
+            disableClearable
             isOptionEqualToValue={(option: T, value: T) => {
                 if (!value) return false;
                 return option === value;
@@ -34,10 +49,17 @@ export const FinalFormAutocomplete = <
                     {...params}
                     {...params.InputProps}
                     endAdornment={
-                        <React.Fragment>
-                            {loading ? <CircularProgress color="inherit" size={16} /> : null}
-                            {params.InputProps.endAdornment}
-                        </React.Fragment>
+                        loading || clearable ? (
+                            <>
+                                {loading && <CircularProgress color="inherit" size={16} />}
+                                {clearable && (
+                                    <ClearInputAdornment position="end" hasClearableContent={Boolean(value)} onClick={() => onChange("")} />
+                                )}
+                                {params.InputProps.endAdornment}
+                            </>
+                        ) : (
+                            params.InputProps.endAdornment
+                        )
                     }
                 />
             )}