+++ packages/admin/admin/src/form/Select.tsx
@@ -1,4 +1,17 @@
-import { InputBase, Select as MuiSelect, SelectProps } from "@mui/material";
+import { ChevronDown } from "@comet/admin-icons";
+import { InputAdornment, InputBase, Select as MuiSelect, SelectProps } from "@mui/material";
 import * as React from "react";
 
-export const Select: React.FunctionComponent<SelectProps> = (props) => <MuiSelect input={<InputBase />} {...props} />;
+type IconProps = {
+    className: string;
+};
+
+const Icon = ({ className }: IconProps) => (
+    <InputAdornment position="end" className={className}>
+        <ChevronDown />
+    </InputAdornment>
+);
+
+export const Select = ({ IconComponent = Icon, ...restProps }: SelectProps): React.ReactElement => {
+    return <MuiSelect IconComponent={IconComponent} input={<InputBase />} {...restProps} />;
+};